[XLA:GPU] Rename command.h/cc to command_thunk.h/cc
Rename files and BUILD target to match the class rename of Command →
CommandThunk.
