+++ xla/backends/gpu/profiler/BUILD
@@ -75,7 +75,7 @@ xla_test(
         ":kernel_name_tracer",
         "//xla:shape_util",
         "//xla:xla_data_proto_cc",
-        "//xla/backends/gpu/runtime:command",
+        "//xla/backends/gpu/runtime:command_thunk",
         "//xla/backends/gpu/runtime:command_buffer_cmd",
         "//xla/backends/gpu/runtime:command_buffer_thunk",
         "//xla/backends/gpu/runtime:command_executor",
